requirements.txt
langchain
gradio
flast 
streamlit
requests
openai
python-dotenv
transformers